show number of upvotes and dowwnvotes

+++ app/models/vote.rb
@@ -7,7 +7,7 @@ class Vote < ActiveRecord::Base
   private
 
   def destroy_previous_vote
-    return if !self.votable.post.has_single_picture?
+    return if !self.votable.post.has_single_picture? or self.votable.is_a?(Comment)
     previous_vote = self.votable.votes.where(user_id: self.user_id).take
     previous_vote.destroy if previous_vote != self
   end
